extract modals for left sidebar into own modules
